pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "safetycompanion-rag"
version = "0.1.0"
description = "Structure-first RAG pipeline for SafetyCompanion 2026"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "pandas",
  "pydantic>=2",
  "pypdf",
  "pdfplumber",
  "PyYAML",
  "rank-bm25",
  "scikit-learn"
]

[project.optional-dependencies]
dev = [
  "pytest"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
